update with integration test
