+++ .env.test
@@ -1,0 +1,8 @@
+TEST_DB_NAME=vehicle_db_test
+DB_USER=postgres
+DB_PASSWORD=jackkevin
+DB_HOST=localhost
+DB_PORT=5432
+NODE_ENV=test
+ACCESS_TOKEN_SECRET=secret1234567890
+REFRESH_TOKEN_SECRET=secret0987654321
