+++ config/db.js
@@ -1,8 +1,14 @@
+//config/db.js
 import { Sequelize } from 'sequelize';
 import logger from '../middleware/logger.js';
 
+const isTestEnv = process.env.NODE_ENV === 'test';
+
+// Configure database connection based on environment
 const sequelize = new Sequelize(
-  process.env.DB_NAME || 'vehicle_db',
+  process.env.NODE_ENV === 'test' 
+    ? process.env.TEST_DB_NAME || 'vehicle_db_test'
+    : process.env.DB_NAME || 'vehicle_db',
   process.env.DB_USER || 'postgres',
   process.env.DB_PASSWORD || 'jackkevin',
   {
@@ -16,8 +22,8 @@ const sequelize = new Sequelize(
 export const connectDB = async () => {
   try {
     await sequelize.authenticate();
-    await sequelize.sync(); // Creates tables if they don't exist
-    logger.info('PostgreSQL Connected');
+    await sequelize.sync({ force: isTestEnv }); // Force sync in test env to reset DB
+    logger.info(`PostgreSQL Connected (${isTestEnv ? 'Test' : 'Production'} Environment)`);
   } catch (error) {
     logger.error('PostgreSQL Connection Error:', error);
     process.exit(1);
