+++ controllers/vehicle.controller.js
@@ -1,3 +1,4 @@
+//controllers/vehicle.controller.js
 import Vehicle from '../models/vehicle.model.js';
 import { Op } from 'sequelize';
 import logger from '../middleware/logger.js';
