+++ index.js
@@ -1,3 +1,4 @@
+//index.js
 import { app } from './app.js';
 import logger from './middleware/logger.js';
