+++ middleware/logger.js
@@ -11,3 +11,4 @@ const logger = pino({
 });
 
 export default logger;
+
