+++ package.json
@@ -23,12 +23,16 @@
     "express": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
     "nodemon": "^3.1.9",
+    "pg": "^8.16.0",
     "pino": "^9.6.0",
-    "pino-http": "^10.4.0",
+    "pino-http": "^10.5.0",
     "pino-pretty": "^13.0.0",
+    "sequelize": "^6.37.7",
     "uuid": "^11.1.0"
   },
   "devDependencies": {
-    "@vitest/coverage-v8": "^3.1.2"
+    "@vitest/coverage-v8": "^3.1.2",
+    "jest": "^29.7.0",
+    "supertest": "^7.1.1"
   }
 }
