+++ routes/auth.routes.js
@@ -1,3 +1,4 @@
+//routes/auth.routes.js
 import { Router } from 'express';
 import {
   register,
