+++ setupTest/setupTests.js
@@ -1,0 +1,5 @@
+import { initializeTestDatabase } from '../utils/initializeTestDatabase.js';
+
+beforeAll(async () => {
+  await initializeTestDatabase();
+});
