+++ tests/setup.js
@@ -1,0 +1,2 @@
+import { config } from 'dotenv';
+config({ path: '.env.test' });
